Hide facebook share box if facebook is not enabled

+++ templates/foodleresponse.php
@@ -76,12 +76,13 @@ function show_response($response) {
 	<?php 
 	
 	
-	echo '<div style="" id="facebookshare" title="' . $this->t('facebookshareheader'). '">';
-	echo '<p>' . $this->t('facebooklinkabout') . '<br /><input type="text" style="width: 90%" name="furl" value="' . htmlentities($this->data['url']) . '&amp;auth=facebook" /></p>';
-	echo '<p><a class="button" style="display: block" href="http://www.facebook.com/sharer.php?u=' . urlencode($this->data['url'] . '&amp;auth=facebook') . '&amp;t=' . urlencode('Foodle: ' . $this->data['header']) . '">' . 
-			'<span>' . $this->t('linkonfacebook') . '</span></a></p>';
-	echo '</div>';	
-
+	if (array_key_exists('facebookshare', $this->data) && $this->data['facebookshare']) {
+		echo '<div style="" id="facebookshare" title="' . $this->t('facebookshareheader'). '">';
+		echo '<p>' . $this->t('facebooklinkabout') . '<br /><input type="text" style="width: 90%" name="furl" value="' . htmlentities($this->data['url']) . '&amp;auth=facebook" /></p>';
+		echo '<p><a class="button" style="display: block" href="http://www.facebook.com/sharer.php?u=' . urlencode($this->data['url'] . '&amp;auth=facebook') . '&amp;t=' . urlencode('Foodle: ' . $this->data['header']) . '">' . 
+				'<span>' . $this->t('linkonfacebook') . '</span></a></p>';
+		echo '</div>';	
+	}
 
 
 	$editlocked = FALSE;
